requirements.txt
feedparser==6.0.10
requests==2.31.0
beautifulsoup4==4.12.2
openai>=1.55.3
schedule==1.2.1
python-dateutil==2.8.2
nltk==3.8.1
scikit-learn==1.3.2
httpx<0.28
tqdm>=4.65.0
colorama>=0.4.6 